feat: Do not fail in case gpiochip isn't 0 or 4
In the scenario where user should be abble to use gpio-mocking locally, doesn't make sense to fail if gpiochip is different from 0 or 4.
Because if the user's machine already have an /dev/gpiochip* (or more than one) there is a high chance of gpiochip number for the gpio-mockup will be different from 0 and 4.

Signed-off-by: Lincoln Wallace <lincoln.wallace@canonical.com>